+++ .github/workflows/ascend_accuracy_ci.yml
@@ -47,14 +47,17 @@ jobs:
           pip install -e .\[npu\] --index-url https://bytedpypi.byted.org/simple
           python3 -m build .
           pip install dist/*.whl --force-reinstall --no-deps
-
       - name: test-Base
         shell: bash
         run: |
           echo "Running base tests..."
           pip install transformers==5.1.0
           python3 -m pytest tests/base
-    
+      - name: test-distributed
+        shell: bash
+        run: |
+          echo "Running distributed tests..."
+          ASCEND_RT_VISIBLE_DEVICES=0,1 torchrun --nproc_per_node=2 -m pytest -q tests/distributed/test_expert_parallel.py 
       - name: test-TTX
         shell: bash
         run: |
