+++ mojo_opset/backends/ttx/kernels/npu/flash_attention.py
@@ -7,6 +7,7 @@ import triton
 import triton.language as tl
 
 from mojo_opset.backends.ttx.kernels.npu.utils import get_num_cores
+from mojo_opset.backends.ttx.kernels.utils import prepare_chunk_indices
 
 
 @triton.jit
