+++ mojo_opset/backends/ttx/kernels/npu/group_gemm.py
@@ -2,8 +2,6 @@ import torch
 import triton
 import triton.language as tl
 
-from torch import Tensor
-
 from mojo_opset.backends.ttx.kernels.npu.utils import get_num_cores
