+++ mojo_opset/core/operators/gemm.py
@@ -10,13 +10,11 @@ from ..operator import MojoOperator
 class MojoGroupGemm(MojoOperator):
     def __init__(
         self,
-        group_num,
-        in_feature,
-        out_feature,
+        weight,
         trans_weight=False,
     ):
         super().__init__()
-        self.weight = torch.nn.Parameter(torch.empty(group_num, out_feature, in_feature, **self.tensor_factory_kwargs))
+        self.weight = weight
         self.trans_weight = trans_weight
 
     def forward(self, input: torch.Tensor, group_list: torch.Tensor) -> torch.Tensor:
@@ -40,19 +38,22 @@ class MojoGroupGemm(MojoOperator):
             - Each group's output is computed as `input_g @ weight_g`.
         """
 
+        if isinstance(input, DTensor):
+            input = input.to_local()
+
         if isinstance(self.weight, DTensor):
             weight = self.weight.to_local()
         else:
             weight = self.weight
 
-        if isinstance(input, DTensor):
-            input = input.to_local()
+        if group_list.device.type != "cpu":
+            group_list = group_list.to("cpu")
 
         assert input.dim() == 2, "input must be 2D"
         assert weight.dim() == 3, "weight must be 3D"
 
         num_groups = group_list.numel()
-        assert weight.size(0) == num_groups, "self.weight must have same group count as group_list"
+        assert weight.size(0) == num_groups, "weight group count must match group_list length"
 
         if self.trans_weight:
             weight = weight.transpose(1, 2).contiguous()
