+++ mojo_opset/distributed/parallel/__init__.py
@@ -1,3 +1,1 @@
-from .mojo_style import MojoColwiseParallel
-from .mojo_style import MojoRowwiseParallel
-from .mojo_style import MojoExpertParallel
+from .expert_parallel import MojoExpertParallel
