+++ mojo_opset/modeling/inference_demo.py
@@ -1,11 +1,14 @@
 import argparse
-import os
-import torch
-import json
 import importlib
-from transformers import AutoTokenizer
-from mojo_opset.utils.hf_utils import build_model_from_hf, _resolve_local_files_only
+import json
+import os
 
+import torch
+
+from transformers import AutoTokenizer
+
+from mojo_opset.utils.hf_utils import _resolve_local_files_only
+from mojo_opset.utils.hf_utils import build_model_from_hf
 
 ARCH_MAP = {
     "Qwen3ForCausalLM": ("mojo_qwen3_dense", "Qwen3ForCausalLM"),
