+++ tests/base/test_assertion_rewrite.py
@@ -1,6 +1,6 @@
-import os
 import pytest
 import torch
+
 from mojo_opset.core.operators.kv_cache import MojoStorePagedKVCache
